create empty polygon + reorganize code tests types
